double_tree + prim: Small changes
- Changed the output sentence in double_tree
- Changed nodes_counter to int in prim

+++ prim.cpp
@@ -18,7 +18,7 @@ multimap<Edge *, double> Prim::perform_prim(double start_node_)
 
     double total_MST_weight = 0.;
 
-    double nodes_counter = _graph->get_nodes().size();
+    int nodes_counter = _graph->get_nodes().size();
 
 //    _MST_graph = new Graph (_graph->is_weighted(),_graph->is_directed());
